# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BiasTileSetItem": {
        "additionalProperties": false,
        "properties": {
          "formats": {
            "items": {
              "type": "string"
            },
            "title": "Formats",
            "type": "array"
          },
          "layer": {
            "title": "Layer",
            "type": "string"
          },
          "level_key": {
            "title": "Level Key",
            "type": "string"
          },
          "max_zoom": {
            "title": "Max Zoom",
            "type": "integer"
          },
          "min_zoom": {
            "title": "Min Zoom",
            "type": "integer"
          },
          "tile": {
            "$ref": "#/components/schemas/TileTemplate"
          },
          "time_key": {
            "title": "Time Key",
            "type": "string"
          }
        },
        "required": [
          "layer",
          "time_key",
          "level_key",
          "min_zoom",
          "max_zoom",
          "tile"
        ],
        "title": "BiasTileSetItem",
        "type": "object"
      },
      "BiasTileSetsResponse": {
        "additionalProperties": false,
        "properties": {
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BiasTileSetItem"
            },
            "title": "Items",
            "type": "array"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "title": "BiasTileSetsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoricalStatisticItem": {
        "additionalProperties": false,
        "properties": {
          "dataset_path": {
            "title": "Dataset Path",
            "type": "string"
          },
          "metadata_path": {
            "title": "Metadata Path",
            "type": "string"
          },
          "samples": {
            "title": "Samples",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "tiles": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TileTemplate"
            },
            "title": "Tiles",
            "type": "object"
          },
          "variable": {
            "title": "Variable",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "window_end": {
            "title": "Window End",
            "type": "string"
          },
          "window_key": {
            "title": "Window Key",
            "type": "string"
          },
          "window_kind": {
            "title": "Window Kind",
            "type": "string"
          },
          "window_start": {
            "title": "Window Start",
            "type": "string"
          }
        },
        "required": [
          "source",
          "variable",
          "window_kind",
          "window_key",
          "version",
          "window_start",
          "window_end",
          "samples",
          "dataset_path",
          "metadata_path"
        ],
        "title": "HistoricalStatisticItem",
        "type": "object"
      },
      "HistoricalStatisticsResponse": {
        "additionalProperties": false,
        "properties": {
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoricalStatisticItem"
            },
            "title": "Items",
            "type": "array"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "title": "HistoricalStatisticsResponse",
        "type": "object"
      },
...
      "TileTemplate": {
        "additionalProperties": false,
        "properties": {
          "legend": {
            "title": "Legend",
            "type": "string"
          },
          "template": {
            "title": "Template",
            "type": "string"
          }
        },
        "required": [
          "template",
          "legend"
        ],
        "title": "TileTemplate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Digital Earth API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analytics/bias/tile-sets": {
      "get": {
        "operationId": "list_bias_tile_sets_api_v1_analytics_bias_tile_sets_get",
        "parameters": [
          {
            "in": "query",
            "name": "layer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Layer"
            }
          },
          {
            "in": "query",
            "name": "fmt",
            "required": false,
            "schema": {
              "default": "png",
              "title": "Fmt",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BiasTileSetsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Bias Tile Sets",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/historical/statistics": {
      "get": {
        "operationId": "list_historical_statistics_api_v1_analytics_historical_statistics_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "variable",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Variable"
            }
          },
          {
            "in": "query",
            "name": "window_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Window Kind"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "in": "query",
            "name": "fmt",
            "required": false,
            "schema": {
              "default": "png",
              "title": "Fmt",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalStatisticsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Historical Statistics",
        "tags": [
          "analytics"
        ]
      }
    },
...
  }
}